small changes
took the count++ from the while loop and moved it inside the if statement that way it only increments the count variable if the sequence number from the client matches the expected received sequence number. also added a sleep(1000) (sleep for 1 second ) in the client to try and test timeout.